requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
google-cloud-storage==2.18.2
librosa==0.10.1
soundfile==0.12.1
scikit-learn==1.3.2
numpy==1.24.3
PyJWT==2.8.0
bcrypt==4.1.2
httpx==0.27.0
pydub==0.25.1
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
aiofiles==24.1.0
openai==1.12.0
pytz==2024.1

# 백그라운드 작업 큐 (Celery + Redis)
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Notion 연동
notion-client==2.2.1
cryptography==42.0.5

# AWS S3 (파일 저장)
boto3==1.35.0

# FCM 푸시 알림
pyfcm==2.1.0

# 고급 오디오 처리
noisereduce==3.0.3
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4
numpy==1.24.3

